final printf (manque Makefile)

+++ ft_count.c
@@ -6,7 +6,7 @@
 /*   By: tmanolis <tmanolis@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/06 18:31:10 by tmanolis          #+#    #+#             */
-/*   Updated: 2021/09/06 18:31:59 by tmanolis         ###   ########.fr       */
+/*   Updated: 2021/09/07 14:54:13 by tmanolis         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,7 +47,7 @@ size_t		ft_count_decimals_unsigned(unsigned int nb)
 	return (count);
 }
 
-size_t		ft_count_hexa(unsigned int nb)
+size_t		ft_count_hexa(unsigned long int nb)
 {
 	size_t count;
